+++ tests/test_processing.py
@@ -1986,6 +1986,8 @@ class TestNoise(unittest.TestCase):
         self.processing = aspecd.processing.Noise()
         self.dataset = aspecd.dataset.Dataset()
         self.dataset.data.data = np.zeros(2**12)
+        self.dataset2d = aspecd.dataset.Dataset()
+        self.dataset2d.data.data = np.zeros([2**12, 2**12])
 
     @staticmethod
     def slope_of_power_spectral_density(noise):
@@ -2042,3 +2044,17 @@ class TestNoise(unittest.TestCase):
         self.dataset.process(self.processing)
         self.assertAlmostEqual(-1., self.slope_of_power_spectral_density(
             self.dataset.data.data), delta=0.1)
+
+    @unittest.skip
+    def test_2d_dataset(self):
+        self.dataset2d.process(self.processing)
+        self.assertTrue(self.dataset2d.data.data.all())
+
+    @unittest.skip
+    def test_2d_dataset_with_pink_noise_is_pink_in_first_dimension_only(self):
+        self.processing.parameters['exponent'] = -1
+        self.dataset2d.process(self.processing)
+        self.assertAlmostEqual(-1., self.slope_of_power_spectral_density(
+            self.dataset2d.data.data[0, :]), delta=0.1)
+        self.assertAlmostEqual(0., self.slope_of_power_spectral_density(
+            self.dataset2d.data.data[:, 0]), delta=0.1)